tools: wm: - cast size_t to int
To prevent build failure during building with Hardening options enabled due to "format '%d' expects argument of type 'int', but argument 5 has type 'size_t'", the fifth argument is cast to integer on the input.
Author: Paul Gevers <elbrus@debian.org>

+++ tools/wml/wmloutkey.c
@@ -618,10 +618,10 @@ void wmlOutputUilKeyTabBody (outfil, tokvec, maxlen, maxkey)
         wmlTokenClassString (tkclass, tok);
         wmlTokenSymKString (tksym, tok);
         wmlTokenTokenString (tktoken, tok);
-        fprintf (outfil, "    {%s, %s, %ld, %s, \"%s\"},\n",
+        fprintf (outfil, "    {%s, %s, %d, %s, \"%s\"},\n",
             tkclass,
             tksym,
-            strlen(tokstg),
+            (int)strlen(tokstg),
             tktoken,
             tokstg);
         if ( (int)strlen(tokstg) > *maxlen )
